+++ Sources/Danger/DangerDSL.swift
@@ -16,7 +16,9 @@ public struct DangerDSL: Decodable {
     public let github: GitHub!
 
     public let bitbucket_server: BitBucketServer!
-    
+
+    public let utils: DangerUtils!
+
     enum CodingKeys: String, CodingKey {
         case git
         case github
@@ -29,7 +31,8 @@ public struct DangerDSL: Decodable {
         self.git = try container.decode(Git.self, forKey: .git)
         self.github = try container.decodeIfPresent(GitHub.self, forKey: .github)
         self.bitbucket_server = try container.decodeIfPresent(BitBucketServer.self, forKey: .bitbucket_server)
-        
+        self.utils = DangerUtils()
+
         let settings = try container.decode(Settings.self, forKey: .settings)
         
         if runningOnGithub {